Drop charset conversation in as_string for the moment as it's broken
Assume you import an UTF-8 and quoted-printable encoded vCard. During
export via as_string it gets double-encoded which breaks the exported
one.

Charset conversation is not trivial as there might be "not-map-able"
characters from original charset to the desired target charset. So
lets drop this feature for the moment and re-start later with a
clean approach.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>